refactor(service): extract GroupExecutor + mutationLocation utilities
MutationTestingService had grown into a 800+ line orchestrator class
mixing lifecycle (setup, baseline, plan, loop, rollback) with
per-iteration evaluation (deploy, run, classify, build mutant result).
Two extractions to separate those responsibilities:

- src/service/groupExecutor.ts (new): owns the per-iteration evaluation.
  Holds session-scoped collaborators (apexClass, tokenStream,
  testMethodsPerLine, generator, repo, runner, progress, messages) and
  exposes a single evaluate(group, completedSoFar, loopStartTime,
  totalMutations) method. Internally handles announce → deploy → run →
  attribute → progress-update. classifyError, hasCoverageGap, and
  buildMutantResult live here as private helpers.

- src/service/mutationLocation.ts (new): pure free functions
  calculateMutationPosition and extractMutationOriginalText shared by
  the executor (for runtime mutant results) and the service (for
  dry-run output). advancePosition is a private helper.

- src/service/timeUtils.ts: hosts formatRemainingTime (moved from the
  service) alongside the existing formatDuration / timeExecution.

MutationTestingService shrinks from ~870 to 559 lines and now reads as
a lifecycle orchestrator: createAdapters, fetchApexClass, discoverTypes,
verify*, runBaselineTests, generateMutations, planGroups,
executeMutationLoop (~30 lines, just iterates groups and delegates to
the executor), rollback, calculateScore, displayTimeEstimate,
filterTestMethods, buildDryRunResult, buildMutatorFilter,
extractCoveredLines.

Tests for the moved free functions migrated to dedicated files
(mutationLocation.test.ts new, formatRemainingTime cases moved to
timeUtils.test.ts). All 1548 unit tests pass at 100% branch coverage.
